Refactor: package classes
Packaged classes according to service, control, util packages
