+++ src/main/java/iudx/data/marketplace/authenticator/AuthenticationVerticle.java
@@ -1,5 +1,7 @@
 package iudx.data.marketplace.authenticator;
 
+import iudx.data.marketplace.authenticator.service.AuthenticationService;
+import iudx.data.marketplace.authenticator.service.JwtAuthenticationServiceImpl;
 import static iudx.data.marketplace.authenticator.util.Constants.AUTH_JWKS_PATH;
 import static iudx.data.marketplace.common.Constants.AUTH_SERVICE_ADDRESS;
 import static iudx.data.marketplace.common.Constants.JWT_LEEWAY_TIME;
@@ -26,7 +28,7 @@ import org.apache.logging.log4j.Logger;
  * <h1>Authentication Verticle</h1>
  *
  * <p>The Authentication Verticle implementation in the IUDX ACL-APD Server exposes the
- * {@link iudx.data.marketplace.authenticator.AuthenticationService } over the Vert.x Event Bus.
+ * {@link AuthenticationService } over the Vert.x Event Bus.
  *
  * @version 1.0
  * @since 2020-05-31
