+++ src/main/java/iudx/data/marketplace/authenticator/handlers/authentication/AuthHandler.java
@@ -7,8 +7,8 @@ import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.ext.web.RoutingContext;
 import iudx.data.marketplace.apiserver.exceptions.DxRuntimeException;
-import iudx.data.marketplace.authenticator.AuthenticationService;
-import iudx.data.marketplace.authenticator.model.JwtData;
+import iudx.data.marketplace.authenticator.service.AuthenticationService;
+import iudx.data.marketplace.authenticator.service.model.JwtData;
 import iudx.data.marketplace.common.HttpStatusCode;
 import iudx.data.marketplace.common.RoutingContextHelper;
 import org.apache.logging.log4j.LogManager;
