+++ src/main/java/iudx/data/marketplace/authenticator/handlers/authorization/AuthorizationHandler.java
@@ -5,9 +5,8 @@ import static iudx.data.marketplace.common.HttpStatusCode.UNAUTHORIZED;
 import io.vertx.core.Handler;
 import io.vertx.ext.web.RoutingContext;
 import iudx.data.marketplace.apiserver.exceptions.DxRuntimeException;
-import iudx.data.marketplace.authenticator.model.DxRole;
-import iudx.data.marketplace.authenticator.model.JwtData;
-import iudx.data.marketplace.common.HttpStatusCode;
+import iudx.data.marketplace.authenticator.service.model.DxRole;
+import iudx.data.marketplace.authenticator.service.model.JwtData;
 import iudx.data.marketplace.common.ResponseUrn;
 import iudx.data.marketplace.common.RoutingContextHelper;
 import java.util.Arrays;
