+++ src/main/java/iudx/data/marketplace/authenticator/service/package-info.java
@@ -2,6 +2,6 @@
     groupPackage = "iudx.data.marketplace.authenticator",
     name = "iudx-data-marketplace-authentication-service",
     useFutures = true)
-package iudx.data.marketplace.authenticator;
+package iudx.data.marketplace.authenticator.service;
 
 import io.vertx.codegen.annotations.ModuleGen;
