+++ src/test/java/iudx/data/marketplace/apiserver/handlers/AuthHandlerTest.java
@@ -14,7 +14,7 @@ import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
 import iudx.data.marketplace.Util;
 import iudx.data.marketplace.aaaService.AuthClient;
-import iudx.data.marketplace.authenticator.AuthenticationService;
+import iudx.data.marketplace.authenticator.service.AuthenticationService;
 import iudx.data.marketplace.authenticator.handlers.authentication.AuthHandler;
 import iudx.data.marketplace.common.Api;
 import iudx.data.marketplace.policies.service.model.User;
