+++ src/test/java/iudx/data/marketplace/apiserver/handlers/ExceptionHandlerTest.java
@@ -8,7 +8,7 @@ import io.vertx.ext.web.RoutingContext;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
 import iudx.data.marketplace.apiserver.exceptions.DxRuntimeException;
-import iudx.data.marketplace.authenticator.handlers.ExceptionHandler;
+import iudx.data.marketplace.common.ExceptionHandler;
 import iudx.data.marketplace.common.ResponseUrn;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
