+++ src/test/java/iudx/data/marketplace/authenticator/TestAuthClient.java
@@ -4,7 +4,8 @@ import io.vertx.core.buffer.Buffer;
 import iudx.data.marketplace.aaaService.AuthClient;
 import iudx.data.marketplace.apiserver.util.Role;
 import iudx.data.marketplace.authenticator.handlers.authentication.AuthHandler;
-import iudx.data.marketplace.authenticator.model.UserInfo;
+import iudx.data.marketplace.authenticator.service.model.UserInfo;
+import iudx.data.marketplace.authenticator.service.AuthenticationService;
 import iudx.data.marketplace.common.Api;
 import iudx.data.marketplace.policies.service.model.User;
 import org.apache.logging.log4j.LogManager;
