+++ src/test/java/iudx/data/marketplace/authenticator/model/JwtDataTest.java
@@ -3,8 +3,8 @@ package iudx.data.marketplace.authenticator.model;
 import io.vertx.core.json.JsonObject;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
+import iudx.data.marketplace.authenticator.service.model.JwtData;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
